os refactored

+++ os/utils/binary.c
@@ -6,10 +6,10 @@
  */
 
 #include "binary.h"
-#include "address.h"
+#include "../kernel/arch/address.h"
 
 
-static inline uint32_t S_to_binary_(const char *s){
+static inline uint64_t S_to_binary_(const char *s){
         uint32_t i = 0;
         while (*s) {
                 i <<= 1;
